Removed unused syn methods
